# auslib/web/admin/swagger/api.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: "Balrog's Admin API"
  termsOfService: "https://wiki.mozilla.org/Balrog"
  license:
    name: "Mozilla Public License 2.0"
    url: "https://www.mozilla.org/en-US/MPL/2.0/"
  contact:
    name: "Mozilla Release Engineering"
    url: "https://github.com/mozilla/balrog"
    email: "release-engineering@lists.mozilla.org"
  description: >
    **Balrog** is the software that runs the server side component of the update
    system used by Firefox and other Mozilla products. It is the successor to *AUS*
    (Application Update Service), which did not scale to our current needs nor allow
    us to adapt to more recent business requirements. Balrog helps us ship updates faster
    and with much more flexibility than we’ve had in the past.
security:
  - basic: []
securityDefinitions:
  basic:
    type: basic
    description: HTTP basic access authentication

# host and scheme can be omitted for a more dynamic association.
# In this case, the host and scheme used to serve the API documentation
# will be used for API calls. For example, if Swagger UI-based documentation is
# hosted at https://api.example.com/apidocs/index.html, “try it out” API calls
# will be directed to https://api.example.com.

#schemes:
#  - http
#  - https
# basePath: "/api"
# host: "localhost:8080"
# Swagger-UI bug : fix TODO week 13
...
parameters:
...
  allSCParam:
    name: all
    in: query
    description: if argument is present in request then response returns all scheduled changes else only returns unfinished changes
    type: string
    x-nullable: true
    required: false
...
paths:
...
  /scheduled_changes/emergency_shutoff:
    get:
      summary: Returns scheduled changes for emergency shut off.
      description: >
        Returns changes that was automatically scheduled when updates for a product/channel was disabled.
      tags:
        - Scheduled Changes
        - Emergency Shutoff
      operationId: auslib.web.admin.emergency_shutoff.scheduled_changes
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/allSCParam'
      responses:
        '200':
          description: JSON object with scheduled changes for emergency shut off.
          schema:
            type: object
            required:
              - scheduled_changes
            allOf:
              - $ref: "#/definitions/CountModel"
              - type: object
                properties:
                  scheduled_changes:
                    description: Array where each element is a scheduled change for a emergency shut off.
                    type: array
                    minItems: 0
                    items:
                      type: object
                      allOf:
                        - $ref: '#/definitions/SCSignoffsModel'
                        - $ref: '#/definitions/EmergencyShutOffModel'
                        - properties:
                            original_row:
                              $ref: '#/definitions/EmergencyShutOffModel'

    post:
      summary: Create scheduled change for emergency shut off.
      description: >
        Create scheduled change for a emergency shut off. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id34)
      tags:
        - Scheduled Changes
        - Emergency Shutoff
      operationId: auslib.web.admin.emergency_shutoff.schedule_deletion
      consumes:
        - application/json
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object"
        description: "create scheduled change for emergency shut off."
      parameters:
        - name: sc_emergency_shutoff
          in: body
          description: Request body for sc emergency shut off
          required: true
          schema:
            type: object
            allOf:
              - $ref: '#/definitions/CSRFModel'
              - $ref: '#/definitions/SCTime'
              - $ref: '#/definitions/EmergencyShutOffModel'
            properties:
              change_type:
                description: type of change
                type: string
                format: ascii
                enum:
                  - delete
                minLength: 1
                maxLength: 50
                example: update
      responses:
        '200':
          description: Returns sc_id of change
          schema:
            $ref: '#/definitions/SCIdModel'
        '400':
          $ref: '#/responses/invalidFormData'
        '401':
          $ref: '#/responses/unauthenticatedUser'
        '403':
          $ref: '#/responses/unauthorizedUser'
...
definitions:
  CSRFModel:
    title: CSRF Token
    description: CSRF Token in POST/PUT or UPDATE requests
    type: object
    properties:
      csrf_token:
        type: string
        format: ascii
        example: "1491342563##c4e6fef0b978e6c89af9ff1015e67b9ca7c45d14"
...
  SCIdModel:
    title: SC ID
    description: "model to re-use 'sc_id' field"
    type: object
    required:
      - sc_id
    properties:
      sc_id:
        description: id of scheduled change
        type: integer
        format: int32
        minimum: 0
        example: 2
...
  SCSignoffsModel:
    title: SC GET
    description: "model for defining signoffs fields in scheduled changes GET operations"
    type: object
    required:
      - required_signoffs
      - signoffs
    properties:
      required_signoffs:
        type: object
        description: Contains key-value pairs of role names and number of minimum signoffs needed from each role
        example:
          releng: 2
      signoffs:
        type: object
        description: Contains key-value pairs of user name and role name via which the change was signed off
        example:
          releng: 2
...
  SCTime:
    title: Schedule Change Time
    description: "model to re-use 'when' field in  scheduled change."
    type: object
    properties:
      when:
        description: Time (in milliseconds since epoch) when to enact scheduled change
        type: ["integer", "null"]
        format: when
        example: 1499712894
...